requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
DateTime==4.6
DBUtils==3.0.2
Django==4.1.1
fake-useragent==0.1.11
handright==8.0.0
idna==3.3
lxml==4.9.1
mysqlclient==2.1.1
peewee==3.15.2
prettytable==3.3.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pyecharts==1.9.1
PyMySQL==1.0.2
python-highcharts==0.4.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
vincent==0.4.4
wincertstore==0.2
zope.interface==5.4.0
